feat: connect Flight and Spaceship

+++ server/typescript/src/service/spaceship.service.ts
@@ -3,11 +3,15 @@ import { SpaceshipInput, UpdateSpaceshipInput } from "../resolvers/types/spacesh
 
 export default class SpaceshipService {
   async findAll() {
-    return Spaceships.find({}).populate({ path: 'model' });
+    return Spaceships.find({})
+      .populate({ path: 'model' })
+      .populate({ path: 'flights', populate: { path: 'journey' }});
   }
 
   async findById(id: string) {
-    return Spaceships.findById(id).populate({ path: 'model'});
+    return Spaceships.findById(id)
+      .populate({ path: 'model'})
+      .populate({ path: 'flights', populate: { path: 'journey' }});
   }
 
   async addSpaceship(input: SpaceshipInput) {
@@ -29,11 +33,15 @@ export default class SpaceshipService {
   async updateSpaceship(id: string, input: UpdateSpaceshipInput) {
     // * doesn't allow to update model
     const updatedSpaceship = await Spaceships.findByIdAndUpdate(id, input, { new: true });
-    return Spaceships.findById(id).populate({ path: 'model' });
+    return Spaceships.findById(id)
+      .populate({ path: 'model' })
+      .populate({ path: 'flights', populate: { path: 'journey' }});
   }
 
   async deleteSpaceship(id: string) {
     const spaceship = await Spaceships.findById(id);
+
+    if (spaceship.flights.length != 0) return `Spaceship with ID: ${id} is still assigned to flights`;
 
     await Promise.all([
       Spaceships.findByIdAndDelete(id),
